Change KNNMethod to interface
Changes KNNMethod to interface and creates methods per engine/algo
combination. For one, this will make code much more maintainable as we
wont have to deal with big maps and builders. For the other, we can
implement more complex logic between engines.
