+++ src/main/java/org/opensearch/knn/index/engine/lucene/Lucene.java
@@ -7,31 +7,15 @@ package org.opensearch.knn.index.engine.lucene;
 
 import com.google.common.collect.ImmutableMap;
 import org.apache.lucene.util.Version;
-import org.opensearch.knn.common.KNNConstants;
-import org.opensearch.knn.index.KNNSettings;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.engine.JVMLibrary;
 import org.opensearch.knn.index.engine.KNNMethod;
-import org.opensearch.knn.index.engine.MethodComponent;
-import org.opensearch.knn.index.engine.MethodComponentContext;
-import org.opensearch.knn.index.engine.Parameter;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 
-import static org.opensearch.knn.common.KNNConstants.DYNAMIC_CONFIDENCE_INTERVAL;
-import static org.opensearch.knn.common.KNNConstants.ENCODER_SQ;
-import static org.opensearch.knn.common.KNNConstants.LUCENE_SQ_BITS;
-import static org.opensearch.knn.common.KNNConstants.LUCENE_SQ_CONFIDENCE_INTERVAL;
-import static org.opensearch.knn.common.KNNConstants.LUCENE_SQ_DEFAULT_BITS;
-import static org.opensearch.knn.common.KNNConstants.MAXIMUM_CONFIDENCE_INTERVAL;
-import static org.opensearch.knn.common.KNNConstants.METHOD_ENCODER_PARAMETER;
 import static org.opensearch.knn.common.KNNConstants.METHOD_HNSW;
-import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_EF_CONSTRUCTION;
-import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_M;
-import static org.opensearch.knn.common.KNNConstants.MINIMUM_CONFIDENCE_INTERVAL;
 
 /**
  * KNN Library for Lucene
@@ -39,56 +23,8 @@ import static org.opensearch.knn.common.KNNConstants.MINIMUM_CONFIDENCE_INTERVAL
 public class Lucene extends JVMLibrary {
 
     Map<SpaceType, Function<Float, Float>> distanceTransform;
-    private static final List<Integer> LUCENE_SQ_BITS_SUPPORTED = List.of(7);
 
-    private final static MethodComponentContext ENCODER_DEFAULT = new MethodComponentContext(
-        KNNConstants.ENCODER_FLAT,
-        Collections.emptyMap()
-    );
-
-    private final static Map<String, MethodComponent> HNSW_ENCODERS = ImmutableMap.of(
-        ENCODER_SQ,
-        MethodComponent.Builder.builder(ENCODER_SQ)
-            .addParameter(
-                LUCENE_SQ_CONFIDENCE_INTERVAL,
-                new Parameter.DoubleParameter(
-                    LUCENE_SQ_CONFIDENCE_INTERVAL,
-                    null,
-                    v -> v == DYNAMIC_CONFIDENCE_INTERVAL || (v >= MINIMUM_CONFIDENCE_INTERVAL && v <= MAXIMUM_CONFIDENCE_INTERVAL)
-                )
-            )
-            .addParameter(
-                LUCENE_SQ_BITS,
-                new Parameter.IntegerParameter(LUCENE_SQ_BITS, LUCENE_SQ_DEFAULT_BITS, LUCENE_SQ_BITS_SUPPORTED::contains)
-            )
-            .build()
-    );
-
-    final static Map<String, KNNMethod> METHODS = ImmutableMap.of(
-        METHOD_HNSW,
-        KNNMethod.Builder.builder(
-            MethodComponent.Builder.builder(METHOD_HNSW)
-                .addParameter(
-                    METHOD_PARAMETER_M,
-                    new Parameter.IntegerParameter(METHOD_PARAMETER_M, KNNSettings.INDEX_KNN_DEFAULT_ALGO_PARAM_M, v -> v > 0)
-                )
-                .addParameter(
-                    METHOD_PARAMETER_EF_CONSTRUCTION,
-                    new Parameter.IntegerParameter(
-                        METHOD_PARAMETER_EF_CONSTRUCTION,
-                        KNNSettings.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION,
-                        v -> v > 0
-                    )
-                )
-                .addParameter(
-                    METHOD_ENCODER_PARAMETER,
-                    new Parameter.MethodComponentContextParameter(METHOD_ENCODER_PARAMETER, ENCODER_DEFAULT, HNSW_ENCODERS)
-                )
-                .build(),
-            new LuceneHNSWContext()
-
-        ).addSpaces(SpaceType.UNDEFINED, SpaceType.L2, SpaceType.COSINESIMIL, SpaceType.INNER_PRODUCT).build()
-    );
+    final static Map<String, KNNMethod> METHODS = ImmutableMap.of(METHOD_HNSW, new LuceneHNSWMethod());
 
     // Map that overrides the default distance translations for Lucene, check more details in knn documentation:
     // https://opensearch.org/docs/latest/search-plugins/knn/approximate-knn/#spaces
