+++ src/main/java/org/opensearch/knn/index/engine/nmslib/NmslibHNSWMethod.java
@@ -1,0 +1,61 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.opensearch.knn.index.engine.nmslib;
+
+import org.opensearch.knn.index.KNNSettings;
+import org.opensearch.knn.index.SpaceType;
+import org.opensearch.knn.index.engine.AbstractKNNMethod;
+import org.opensearch.knn.index.engine.DefaultHnswContext;
+import org.opensearch.knn.index.engine.MethodComponent;
+import org.opensearch.knn.index.engine.Parameter;
+
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
+
+import static org.opensearch.knn.common.KNNConstants.METHOD_HNSW;
+import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_EF_CONSTRUCTION;
+import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_M;
+
+/**
+ * Nmslib's HNSW implementation
+ */
+public class NmslibHNSWMethod extends AbstractKNNMethod {
+
+    public final static List<SpaceType> SUPPORTED_SPACES = Arrays.asList(
+        SpaceType.UNDEFINED,
+        SpaceType.L2,
+        SpaceType.L1,
+        SpaceType.LINF,
+        SpaceType.COSINESIMIL,
+        SpaceType.INNER_PRODUCT
+    );
+
+    /**
+     * Constructor. Builds the method with the default parameters and supported spaces.
+     * @see AbstractKNNMethod
+     */
+    public NmslibHNSWMethod() {
+        super(initMethodComponent(), Set.copyOf(SUPPORTED_SPACES), new DefaultHnswContext());
+    }
+
+    private static MethodComponent initMethodComponent() {
+        return MethodComponent.Builder.builder(METHOD_HNSW)
+            .addParameter(
+                METHOD_PARAMETER_M,
+                new Parameter.IntegerParameter(METHOD_PARAMETER_M, KNNSettings.INDEX_KNN_DEFAULT_ALGO_PARAM_M, v -> v > 0)
+            )
+            .addParameter(
+                METHOD_PARAMETER_EF_CONSTRUCTION,
+                new Parameter.IntegerParameter(
+                    METHOD_PARAMETER_EF_CONSTRUCTION,
+                    KNNSettings.INDEX_KNN_DEFAULT_ALGO_PARAM_EF_CONSTRUCTION,
+                    v -> v > 0
+                )
+            )
+            .build();
+    }
+}
