+++ src/test/java/org/opensearch/knn/index/engine/AbstractKNNMethodTests.java
@@ -16,21 +16,22 @@ import org.opensearch.knn.training.VectorSpaceInfo;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 
 import static org.opensearch.knn.common.KNNConstants.NAME;
 import static org.opensearch.knn.common.KNNConstants.PARAMETERS;
 import static org.opensearch.knn.common.KNNConstants.METHOD_PARAMETER_SPACE_TYPE;
 
-public class KNNMethodTests extends KNNTestCase {
+public class AbstractKNNMethodTests extends KNNTestCase {
 
-    /**
-     * Test KNNMethod method component getter
-     */
-    public void testGetMethodComponent() {
-        String name = "test";
-        KNNMethod knnMethod = KNNMethod.Builder.builder(MethodComponent.Builder.builder(name).build(), EMPTY_ENGINE_SPECIFIC_CONTEXT)
-            .build();
-        assertEquals(name, knnMethod.getMethodComponent().getName());
+    private static class TestKNNMethod extends AbstractKNNMethod {
+        public TestKNNMethod(
+            MethodComponent methodComponent,
+            Set<SpaceType> spaces,
+            EngineSpecificMethodContext engineSpecificMethodContext
+        ) {
+            super(methodComponent, spaces, engineSpecificMethodContext);
+        }
     }
 
     /**
@@ -38,9 +39,11 @@ public class KNNMethodTests extends KNNTestCase {
      */
     public void testHasSpace() {
         String name = "test";
-        KNNMethod knnMethod = KNNMethod.Builder.builder(MethodComponent.Builder.builder(name).build(), EMPTY_ENGINE_SPECIFIC_CONTEXT)
-            .addSpaces(SpaceType.L2, SpaceType.COSINESIMIL)
-            .build();
+        KNNMethod knnMethod = new TestKNNMethod(
+            MethodComponent.Builder.builder(name).build(),
+            Set.of(SpaceType.L2, SpaceType.COSINESIMIL),
+            EMPTY_ENGINE_SPECIFIC_CONTEXT
+        );
         assertTrue(knnMethod.isSpaceTypeSupported(SpaceType.L2));
         assertTrue(knnMethod.isSpaceTypeSupported(SpaceType.COSINESIMIL));
         assertFalse(knnMethod.isSpaceTypeSupported(SpaceType.INNER_PRODUCT));
@@ -51,9 +54,11 @@ public class KNNMethodTests extends KNNTestCase {
      */
     public void testValidate() throws IOException {
         String methodName = "test-method";
-        KNNMethod knnMethod = KNNMethod.Builder.builder(MethodComponent.Builder.builder(methodName).build(), EMPTY_ENGINE_SPECIFIC_CONTEXT)
-            .addSpaces(SpaceType.L2)
-            .build();
+        KNNMethod knnMethod = new TestKNNMethod(
+            MethodComponent.Builder.builder(methodName).build(),
+            Set.of(SpaceType.L2),
+            EMPTY_ENGINE_SPECIFIC_CONTEXT
+        );
 
         // Invalid space
         XContentBuilder xContentBuilder = XContentFactory.jsonBuilder()
@@ -95,9 +100,11 @@ public class KNNMethodTests extends KNNTestCase {
      */
     public void testValidateWithData() throws IOException {
         String methodName = "test-method";
-        KNNMethod knnMethod = KNNMethod.Builder.builder(MethodComponent.Builder.builder(methodName).build(), EMPTY_ENGINE_SPECIFIC_CONTEXT)
-            .addSpaces(SpaceType.L2)
-            .build();
+        KNNMethod knnMethod = new TestKNNMethod(
+            MethodComponent.Builder.builder(methodName).build(),
+            Set.of(SpaceType.L2),
+            EMPTY_ENGINE_SPECIFIC_CONTEXT
+        );
 
         VectorSpaceInfo testVectorSpaceInfo = new VectorSpaceInfo(4);
 
@@ -144,7 +151,7 @@ public class KNNMethodTests extends KNNTestCase {
             .setMapGenerator(((methodComponent1, methodComponentContext) -> methodComponentContext.getParameters()))
             .build();
 
-        KNNMethod knnMethod = KNNMethod.Builder.builder(methodComponent, EMPTY_ENGINE_SPECIFIC_CONTEXT).build();
+        KNNMethod knnMethod = new TestKNNMethod(methodComponent, Set.of(SpaceType.L2), EMPTY_ENGINE_SPECIFIC_CONTEXT);
 
         Map<String, Object> expectedMap = new HashMap<>(generatedMap);
         expectedMap.put(KNNConstants.SPACE_TYPE, spaceType.getValue());
@@ -155,16 +162,14 @@ public class KNNMethodTests extends KNNTestCase {
         );
     }
 
-    public void testBuilder() {
-        String name = "test";
-        KNNMethod.Builder builder = KNNMethod.Builder.builder(MethodComponent.Builder.builder(name).build(), EMPTY_ENGINE_SPECIFIC_CONTEXT);
-        KNNMethod knnMethod = builder.build();
-
-        assertEquals(name, knnMethod.getMethodComponent().getName());
-
-        builder.addSpaces(SpaceType.L2);
-        knnMethod = builder.build();
-
-        assertTrue(knnMethod.isSpaceTypeSupported(SpaceType.L2));
+    public void testGetEngineSpecificMethodContext() {
+        String methodName = "test-method";
+        EngineSpecificMethodContext engineSpecificMethodContext = new DefaultHnswContext();
+        KNNMethod knnMethod = new TestKNNMethod(
+            MethodComponent.Builder.builder(methodName).build(),
+            Set.of(SpaceType.L2),
+            engineSpecificMethodContext
+        );
+        assertEquals(engineSpecificMethodContext, knnMethod.getEngineSpecificMethodContext());
     }
 }
